Refactored to pass test_hadamard
