+++ quantpy/sympy/tests/test_hadamard.py
@@ -1,8 +1,11 @@
 import time
+
+from sympy.physics.quantum.gate import H
 from sympy.physics.quantum.qubit import Qubit
-from sympy.physics.quantum.gate import X,Y,Z,H,T,S,CNOT,SWAP
-from sympy.physics.quantum.gate import IdentityGate as I
+
 from quantpy.sympy.qapply import qapply
+
+
 #from sympy.physics.quantum.qapply import qapply
 
 def test_hadamard_loop():
@@ -28,6 +31,6 @@ def test_hadamard_loop():
             elapsed_time = time.time() - start
             ss += elapsed_time
             print(r)
-            # print ("elapsed_time:\t{0},\t{1} [sec]".format(n,elapsed_time))
-        # print("average:\t{0}x{1},\t{2} [sec]".format(n, lc, ss/lc))
+            print ("elapsed_time:\t{0},\t{1} [sec]".format(n,elapsed_time))
+        print("average:\t{0}x{1},\t{2} [sec]".format(n, lc, ss/lc))
         assert ss/lc < ts
