+++ requirements.txt
@@ -1,1 +1,4 @@
 numpy
+qiskit
+sympy
+IBMQuantumExperience
